+++ src/main/java/com/example/CarArchive/repository/PartRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface PartRepository extends JpaRepository<Part, Integer> {
+public interface PartRepository extends JpaRepository<Part, Long> {
     @Query("SELECT p.name, (p.mileage + p.nextExchange - :inputMileage) AS total FROM Part p WHERE p.nextExchange > 0 AND p.car.id = :inputCarId ORDER BY total ASC")
-    List<Object[]> findPartsByMileage(int inputMileage, int inputCarId);
+    List<Object[]> findPartsByMileage(int inputMileage, Long inputCarId);
 }
