+++ src/main/java/com/example/CarArchive/service/PartService.java
@@ -6,13 +6,13 @@ import java.util.List;
 
 public interface PartService {
     List<Part> getAllParts();
-    Part getPartById(int id);
+    Part getPartById(Long id);
     Part addNewPart(Part part);
-    Part updatePart(int id, Part part);
-    void deletePart(int id);
+    Part updatePart(Long id, Part part);
+    void deletePart(Long id);
 
 
-    List<Object[]> getPartsToExchangeByMileage(int km, int carId);
+    List<Object[]> getPartsToExchangeByMileage(int km, Long carId);
 
 
 }
